Migrate controllers and views to new OBS client
The controllers don't rely on Seeker anymore but instead make use of the
new OBS mini client. Views are adjusted to the small difference between
the old and new data structure.

One example for that is that binary.size does not work as previously, as
size is a built-in method and binary is a modified Hash. Therefore, the
hash-like binary[:size] notation is used.